move to audio mod, use net, and read from channel
not sure about the mod, i'll need to change the network from the process
side i think? so it's gonna be the same name ambeguity mess again, but
that's a problem for tomorrow-amy.
want to figure out how to send whole networks over the channel from the
bevy thread, this would be awesome :3