fix LostWH
it's always applying commands!
running process in post makes sure that when an entity is marked LostWH
it had actually lost the children.
it used to update the net on the frame LostWH is true, but sometimes it
will be marked true, but the children will only be despawned after the
schedule.

before: delete (marks) -> process (looks and finds children) (unmarks) ->
children get despawned after schedule.

now:
- delete runs in update (marks) (children despawned at end of update)
- process finds marked circles that actually have no connections
  (unmarks)

another way would be to sandwich apply_deferred between a chained delete
and process, but that's confusing