contains
not sure about the open checking here. since this will get the array but
only react if the second input is open. this way it works interactively
if the array is fixed and the input number is changing. (just like `get`)

but if it's needed in a case where the array is changing and the number
remains fixed, like in checking if a key is held with `key`'s array..
then you need to keep the wh open with an `open_target`, but then it's
working all the time (updating every frame)

maybe something like hole mirroring would help? if this is open, open
that..