make it a feedback instead (for now)
it has something to do with how we're not processing in blocks.
input signal goes in and pushes output the other end. readhead of the
tap delay is at index 10. all good. but next sample it's at index 11 and
it's reading the number that was at index 10 the previous sample...